stages/subdomain_enum.py
import dns.resolver
import requests
from core.stage import Stage
from core.target import Target
from typing import Set


class SubdomainEnumStage(Stage):
    """Stage 1: Subdomain Enumeration"""
    
    def __init__(self, config: dict):
        super().__init__("Subdomain Enumeration")
        self.config = config
        self.timeout = config.get('timeout', 10)
    
    def execute(self, target: Target) -> bool:
        """Enumerate subdomains for the target"""
        self.notify("start", f"Enumerating subdomains for {target.domain}")
        
        try:
            subdomains = self._enumerate_subdomains(target.domain)
            
            for subdomain in subdomains:
                target.add_subdomain(subdomain)
                self.notify("subdomain_found", subdomain)
            
            self.notify("complete", f"Found {len(subdomains)} subdomains")
            return True
            
        except Exception as e:
            self.notify("error", str(e))
            return False
    
    def _enumerate_subdomains(self, domain: str) -> Set[str]:
        """
        Perform subdomain enumeration using multiple techniques
        Team members should implement:
        1. Dictionary-based brute force
        2. DNS zone transfer
        3. Certificate transparency logs
        4. Search engine queries
        """
        subdomains = set()
        
        # Technique 1: Common subdomain wordlist
        common_subdomains = ['www', 'mail', 'ftp', 'admin', 'blog', 'dev', 'api']
        for sub in common_subdomains:
            subdomain = f"{sub}.{domain}"
            if self._check_subdomain_exists(subdomain):
                subdomains.add(subdomain)
        
        # TODO: Add more enumeration techniques here
        # - crt.sh API for certificate transparency
        # - Subdomain scraping from search engines
        # - DNS brute forcing with larger wordlists
        
        return subdomains
    
    def _check_subdomain_exists(self, subdomain: str) -> bool:
        """Verify if subdomain exists"""
        try:
            dns.resolver.resolve(subdomain, 'A')
            return True
        except:
            return False
